move toolbar comic files in one folder
[tidy] move toolbar comic files in one folder
