+++ client/app/components/toolbar/components/ComicSearch/ComicControls.pcss
@@ -1,4 +1,4 @@
-@import '../../../styles/variables.pcss';
+@import '../../../../styles/variables.pcss';
 
 .comic__controls {
   display: flex;
