+++ client/app/components/toolbar/components/ComicSearch/ComicDetailBack.pcss
@@ -1,0 +1,20 @@
+@import '../../../../styles/variables.pcss';
+
+.comic__back-button {
+  color: var(--colour-white);
+  display: flex;
+  margin-left: auto;
+  align-self: flex-end;
+  background-color: var(--colour-black);
+  position: fixed;
+  right: 0;
+  bottom: 0;
+  cursor: pointer;
+  width: 96px;
+  height: 96px;
+
+  i {
+    width: 100%;
+    height: 100%;
+  }
+}
