+++ client/app/components/toolbar/components/ComicSearch/ComicSearch.pcss
@@ -1,4 +1,4 @@
-@import '../../../styles/variables.pcss';
+@import '../../../../styles/variables.pcss';
 
 :root {
   --tags-element-height: 68px;
