+++ client/app/components/toolbar/components/DetailBack.pcss
@@ -1,6 +1,6 @@
 @import '../../../styles/variables.pcss';
 
-.comic__back-button {
+.detail__back-button {
   color: var(--colour-white);
   display: flex;
   margin-left: auto;
