+++ client/app/components/toolbar/components/Toolbar.jsx
@@ -6,8 +6,8 @@ import pathOr from 'ramda/src/pathOr'
 import any from 'ramda/src/any'
 import {toggleToolbar} from '../actions'
 import classNames from 'classnames'
-import {ComicControls} from './ComicControls'
-import {ComicSearch} from './ComicSearch'
+import {ComicControls} from './ComicSearch/ComicControls'
+import {ComicSearch} from './ComicSearch/ComicSearch'
 import {CreatorControls} from './CreatorSearch/CreatorControls'
 import {CreatorSearch} from './CreatorSearch/CreatorSearch'
 import {CharacterControls} from './CharactersSearch/CharacterControls'
